+++ test/all.dart
@@ -5,6 +5,7 @@ import 'package:test/test.dart';
 
 import 'basic_test.dart' as basic;
 import 'interrupts_test.dart' as interrupts;
+import 'pwm_test.dart' as pwm;
 import 'read_write_test.dart' as read;
 import 'test_util.dart';
 
@@ -17,6 +18,7 @@ runTests() {
   basic.runTests();
   read.runTests();
   interrupts.runTests();
+  pwm.runTests();
 
   test('pins used', () {
     // Ensure LEDs are off