+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## 0.3.0
  * BREAKING CHANGES to match the [fletch gpio package](https://github.com/dart-lang/fletch/blob/master/pkg/gpio/lib/gpio.dart)
  * Renamed PinMode to Mode
- * Rename GpioException GPIOException
+ * Renamed GpioException GPIOException
  * Add Pin.physNum
+ * Add Mode.other to match fletch gpio package
+ * Removed Mode.pulsed in favor of Mode.output and pulseWidth
+ * Removed pwm support for pins other than pin 1
  * Removed top level input, output, pulsed const
  * Removed Gpio.pin method in favor of top level pin function
  * Move classes similar to fletch gpio package into gpio.dart library